Move get_follow_intersection_users
