+++ discovery-provider/src/queries/queries.py
@@ -28,6 +28,7 @@ from src.queries.get_tracks_including_unlisted import get_tracks_including_unlis
 from src.queries.get_stems_of import get_stems_of
 from src.queries.get_feed import get_feed
 from src.queries.get_repost_feed_for_user import get_repost_feed_for_user
+from src.queries.get_follow_intersection_users import get_follow_intersection_users
 
 
 logger = logging.getLogger(__name__)
@@ -107,47 +108,8 @@ def get_repost_feed_for_user_route(user_id):
 # get intersection of users that follow followeeUserId and users that are followed by followerUserId
 # followee = user that is followed; follower = user that follows
 @bp.route("/users/intersection/follow/<int:followee_user_id>/<int:follower_user_id>", methods=("GET",))
-def get_follow_intersection_users(followee_user_id, follower_user_id):
-    users = []
-    db = get_db_read_replica()
-    with db.scoped_session() as session:
-        query = (
-            session.query(User)
-            .filter(
-                User.is_current == True,
-                User.user_id.in_(
-                    session.query(Follow.follower_user_id)
-                    .filter(
-                        Follow.followee_user_id == followee_user_id,
-                        Follow.is_current == True,
-                        Follow.is_delete == False
-                    )
-                    .intersect(
-                        session.query(Follow.followee_user_id)
-                        .filter(
-                            Follow.follower_user_id == follower_user_id,
-                            Follow.is_current == True,
-                            Follow.is_delete == False
-                        )
-                    )
-                )
-            )
-        )
-        users = paginate_query(query).all()
-        users = helpers.query_result_to_list(users)
-        user_ids = [user[response_name_constants.user_id] for user in users]
-
-        current_user_id = get_current_user_id(required=False)
-
-        # bundle peripheral info into user results
-        users = populate_user_metadata(session, user_ids, users, current_user_id)
-
-        # order by follower_count desc
-        users.sort(
-            key=lambda user: user[response_name_constants.follower_count],
-            reverse=True
-        )
-
+def get_follow_intersection_users_route(followee_user_id, follower_user_id):
+    users = get_follow_intersection_users(followee_user_id, follower_user_id)
     return api_helpers.success_response(users)
